requirements.dev.txt
-r requirements.txt

pytest==6.2.2
pytest-django==4.1.0
pre-commit==2.10.1
